// src/app/core/models/piece-type.enum.ts
export enum PieceType {
    // General
    General = 'General',
    // Advisor
    Advisor = 'Advisor',
    // Elephant
    Elephant = 'Elephant',
    // Horse
    Horse = 'Horse',
    // Chariot
    Chariot = 'Chariot',
    // Cannon
    Cannon = 'Cannon',
    // Soldier
    Soldier = 'Soldier',
}